Rename energy_balance_hierarchical_model.* to dynamic_hierarchical_linear_model.*.
