+++ src/HealthGPS/CMakeLists.txt
@@ -26,8 +26,8 @@ target_sources(
             "disease.h"
             "disease_table.cpp"
             "disease_table.h"
-            "energy_balance_hierarchical_model.cpp"
-            "energy_balance_hierarchical_model.h"
+            "dynamic_hierarchical_linear_model.cpp"
+            "dynamic_hierarchical_linear_model.h"
             "energy_balance_model.cpp"
             "energy_balance_model.h"
             "error_message.cpp"
